fix: Account for pagination of results of file system query.
The EFS API for AWS returns paginated results if there are more than the specified # of results (100). This commit updates the query for existing file systems to account for this pagination by looping through the results, and checking if the marker in the respone of the query matches the marker from the previous request. Any new results are added to a slice of file systems which are looped through later to determine which to delete.