test: Raise time to wait for job output readers to start in test.
The test TestServiceGetJobStream_complete is flakey in that sometimes, it takes longer for the job output readers to start, and the job output in the test becomes buffered, which fails the test. This commit doubles the amount of wait time, in an effort to reduce the likelihood that it takes too long for the readers to start to avoid buffering.