server: Be explicit about job id var from original assignment
Prior to this commit, it might of been possible that our job var becomes
nil on error. This means we can no longer "cancel" the job on error
because we've lost the reference to our original job. Instead of
referencing that job.Id, we were already storing it under jobId as a
separate var, so this commit changes that to use the original jobId in
the case where it might of been lost on job errors.