internal/core: Gracefully handle bad proto message from destroyments
Prior to this commit, if a plugin returned a non-Deployment proto
message, we would hard fail the entire operation. Considering that we
are retreiving this message to print what resources were left around or
destroyed, we should not hard fail on this and instead print the error
to the user that we cannot tell them about these resources and move on.
It also incldues the "type" that was returned.