Supporting json configsourcer values
This is a bit hacky, mostly because it continues to use existing hacks. The variable system is requesting it's dynamic default vars as "files". The appconfig system then reads any json-typed variables, dumps the raw json into the "file" contents, and returns it to the variable system.

The variable system then looks at the type the user specified. If it's a string, it'll assume the plugin returned a string value. If it's a bool or number, it'll error. If it's anything else, it'll try reading the plugin output from the "file" as json, and if that works, surface it up the chain as an hcl snippet input variable, which to my great surprise, _just works_.

It's not great that we're guessing at what kind of data the plugins returned based on the type that the user specified, but I'm not up for a broader overhaul of the system at this point to wire more type information through.