bolt: Fix pre-v0.10.4 config sources.
This commit updates the Bolt DB implementation of config sources to "upgrade" a pre-v0.10.4 database record from the config_source table. Prior to v0.10.4, the unique identifier for each record in the table was the name of the plugin. This was changed in v0.10.4 with the addition of scoped config sources, and now the ID is a hashed value, whose hash structure is defined by the plugin name, scope, and workspace name.

This commit therefore deletes any "old" records, those whose ID is the name of any known config sourcer plugin, from Bolt DB, and creates new records with the ID correctly hashed. With this fix, users who had config sources prior to v0.10.3 may continue to interact with those config sources, whereas before this commit, they could not, and the config sources were permanently "stuck" in the database as they were before upgrading.