cli: Fix panic in pipeline inspect, DRY up code
This commit adds some extra guards around when HCP Waypoint returns
empty PipelineRun messages. In this case, we got a PipelineRun message
that's empty with no fields defined, so there should not be any jobs.
This fix guards against that to not panic when attempting to look at the
list of jobs for a given run.

It also simplifies a lot of duplicated code in the inspect command.