bug: Fix key for config sources in BoltDB.
This commit updates how config source records are stored in BoltDB. Previously the key was a uint64 value converted to a byte slice - this is acceptable for BoltDB and reading from it; however, snapshots expect that the database values are UTF-8 encoded, which uint64 is not. This commit therefore converts the uint64 to a string BEFORE converting it to a byte slice, and that is the new "id" for a record in the config_source bucket.