cli: Add `-verbose` flag and rearrange job list columns
This commit hides Application and Pipeline details behind a verbose
flag. These fields are useful, but maybe not all the time at the moment
with Waypoint. They are still visible if the verbose flag is used, but
by default will be hidden to keep the columns shorter and more useful.