cli/runner_install: Do not use cli slice for runner platform
Prior to this commit, the runner install cli attempted to use the target
var for the EnumVar platform. We make this an EnumVar because the flag
package gives good validation when you pass a string that isn't one of
the supported values. In practice however, the CLI only uses a single
value, the first one in the slice. To avoid any panics, we simply use a
string var once the flag package has parsed and actualized its values.

Fixes WAYP-1166