server: Check that UserFromContext gets user before generating token
Prior to this commit, the servers auth would not validate that a proper
user was found for the given context prior to generating a token. We
should not generate a token if no user was found, so this commit updates
the login token func to return an unathenticated error if no user was
found in the given context.

Fixes WAYP-1155