builtin/exec: Properly return a closer for template rendering
Prior to this commit, the closer function for rendering a template was
commented out, meaning that the exec plugin wouldn't properly cleanup
its mess in `/tmp` while rendering commands. This commit updates that
behaviort to return the proper closer func to be called by the caller
once they're finished rendering templates.

Fixes WAYP-1164