internal: Cleanup remote URL trimming for git
Prior to this commit, it was possible to craft a bad git remote and
bypass our normalize func for remote git URLs. TrimLeft and TrimRight
remote all occurances of the given pattern. The intention however is to
only remove the first instance of `git@` and the last instance of `.git`
from the given url. This commit fixes that by updating the string
trimmers to use TrimPrefix and TrimSuffix, which will only remove the
single instance of the pattern.

Fixes WAYP-1173