cli: Ensure that img string passed is proper length
Prior to this commit, if a func calls shortImg with an improper string
length, the whole CLI could panic given the shortImg func has a
hardcoded expectation for the length of the passed in string. This
commit fixes that by first validating the string length and then
returning an error if the passed in string is not long enough to trim as
expected.

Fixes WAYP-1151 (part 1)