plugin/ecs: Add config options for setting the AWS target group protocol and version.
Prior to this commit, only the HTTP protocol was supported. Users of applications which require other protocols, or other protocol versions, such as gRPC, are now supported.

A bug was also fixed, where the user-specified health check protocol was previously being set to the target group protocol, instead of the health check's.