internal/config: Remove redundant context append for parsing
Prior to this commit, if a config had many pipeline stanzas, say more
than 6, each time we would parse a pipeline via the hcl EvalContext we
would append that context to the current one to capture any additional
parts of the config. Given that we are evaluating the entire config file
and not parts of it, this action was redundant. Additionally, it made
parsing many pipeline configs slow for each appended context.

This fixes it by removing the append. We don't need to append this,
given we have the full config context each time we parse pipelines.