fix: Filter global config sources based on workspace scope.
Prior to this commit, global config source get requests were receiving all config sources at the global scope, even if a workspace was specified. The state function was skipping over the filtering out of config sources which were not for the specified workspace. This commit fixes that, and now global config source requests will filter out config sources which are not in the specified workspace.