config_source: Update test, and update error message if none.
This commit updates the test for getting all config sources to verify that the config sources' project, app and workspace are returned, if set.

The CLI command config source-get is also updated here to output a nicer error message if there are no config sources returned when the scope all is used.