feat: Add Ons protobuf updates.
This commit updates the protocol buffer definition to include CRUD for Add Ons and Add On definitions.

This also deprecates a nested proto message TerraformNocodeModule in favor of using a proto message of the same name that isn't nested, as it is commonly used with Add Ons as well as Project Templates.