addons: Explicit fields in Add On creation/update requests.
This commit updates the request proto messages for creating Add Ons, creating Add On definitions, and updating Add On definitions to use explicit individual fields, rather than a whole proto message of an Add On or Add On definition. This makes it clearer for clients which fields they need to provide, and helps avoid setting fields that are managed by the server, such as "id".

AddAddOn is renamed to CreateAddOn.

The type of Add On is refactored out of the AddOn proto message so it can be used by multiple proto messages. It's also renamed to AddOnType.

The ListAddOnsRequest gains a new field, project, so that the returned Add Ons in the response are for the requested project.